+++ Source/Editor/OdysseyVector/Public/OdysseyVectorPath.h
@@ -93,7 +93,7 @@ class ODYSSEYVECTOR_API FOdysseyVectorPath : public FOdysseyVectorObject
         virtual bool PickShape( const ::ULIS::FRectD &iRoi
                               , uint32 iSelectionFlags ) override;
         virtual FOdysseyVectorObject* CopyShape( uint64 iCopyFlags ) override;
-        virtual void Invalidate( uint32 iInvalidationFlags ) override;
+        virtual void Invalidate( uint64 iInvalidationFlags ) override;
         virtual void ExportParam( FOdysseyVectorObject* iDestinationObject
                                 , bool iInvalidate ) override;
